pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rules_generator"
version = "0.0.1"
authors = [
  { name="OpenPecha", email="dev@openpecha.org" },
]
description = "A package to produce word segmentation rules with gold corpus."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
  "botok >= 0.8.12",
  "ordered-set == 4.1.0",
  "PyGithub >= 2.1.1",
]



[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pre-commit",
]


[project.urls]
"Homepage" = "https://github.com/OpenPecha/openpecha-project-template"
"Bug Tracker" = "https://github.com/OpenPecha/openpecha-project-template/issues"
